requirements.txt
django >3.2,<3.3
python-decouple
django-debug-toolbar
djangorestframework
markdown
django-filter